+++ lib/widgets/amount.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/cupertino.dart';
-import 'package:money/helpers/string_helper.dart';
 import 'package:money/models/money_objects/currencies/currency.dart';
 
 /// Formatted text using the supplied currency code and optional the currency/country flag
@@ -17,7 +16,7 @@ class Amount extends StatelessWidget {
   Widget build(final BuildContext context) {
     return Row(
       children: [
-        Text(getCurrencyText(value)),
+        Text(Currency.getCurrencyText(value)),
         const SizedBox(width: 10),
         Currency.buildCurrencyWidget(iso4217),
       ],